Ok.... Basically I've created LevelState.cs which is a Gamestate to hold my terrain and camera and stuff. i suppose it can be TEH GameState.
If you've read the email you'll know that I've messed put everything into a namespace of its own depending on the folder its in. So organise your stuff nicely. I had an issue with Input as a class name and as a namespace. Rashid you can probably fix that.

Apart from that Host Game on the menu will lauch a LevelState.