(Now to actually commit what I was supposed to commit the first time ;) )
Dirty first try on the glue after the decision to give up trying to do things with boost.python.

so far, I am able to get my python code (put into one file) to run from C++ (i.e. The Python interpreter is initialised from a C++ main method and a file with the python script is run.) just like the last commited version of the nicer packaged python code.

In this commit, however, I have already entered the next stage. I am learning how to extract things from python into c++. The goal here is to 1. get the data we need from python and 2., if time permits, to offload as must of the processing to C++ as C++ will be faster (being native and all...).